Simplify cmd/sonobuoy/app/images
- merged images_dynamic.go to images.go
- eventually the delta is just new private functions
