--- cmd/sonobuoy/app/images_dynamic_test.go
@@ -1,1 +1,0 @@
-package app
